+++ modules/consensus/tendermint/primitives/src/lib.rs
@@ -21,4 +21,4 @@ pub use verifier::{
 
 pub mod prover;
 
-pub use prover::ProverError;
+pub use prover::{Client, ProverError};
