+++ tesseract/consensus/clients/src/lib.rs
@@ -1,0 +1,3 @@
+mod polygon;
+
+pub use polygon::HeimdallClient;
